takeFood float
update to make takeFoodCalculation into a float type to make messages still work.

+++ pkg/agents/team3/messages.go
@@ -4,6 +4,7 @@ import (
 	"github.com/SOMAS2021/SOMAS2021/pkg/infra"
 	"github.com/SOMAS2021/SOMAS2021/pkg/messages"
 )
+
 //Upon receipt of message define affected emotions
 // ACK MESSAGE
 //If x time passed no message received/acked morale decrease
@@ -17,7 +18,7 @@ func message(a *CustomAgent3) {
 		a.Log("I got sent a message", infra.Fields{"messageType": receivedMsg.MessageType(), "Food": receivedMsg})
 		if receivedMsg.MessageType() == "askFoodTakenMessage" { //agents response to askFoodTakenMessage
 			a.Log("I've eaten", infra.Fields{"food": takeFoodCalculation(a)})
-			msg := *messages.NewfoodTakenMessage(int(a.Floor()), takeFoodCalculation(a))
+			msg := *messages.NewfoodTakenMessage(int(a.Floor()), float64(takeFoodCalculation(a)))
 			//need to find out how to find where message came from
 			//if receivedMsg.SenderFloor  == a.knowledge.floorBelow{
 			a.SendMessage(-1, msg)
@@ -32,18 +33,18 @@ func message(a *CustomAgent3) {
 		}
 		if receivedMsg.MessageType() == "askIntendedFoodAtmMessage" { //agents response to askIntendedFoodAtmMessage
 			a.Log("My intended food is", infra.Fields{"food": takeFoodCalculation(a)})
-			msg := *messages.NewintendedFoodAtmMessage(int(a.Floor()), takeFoodCalculation(a))
+			msg := *messages.NewintendedFoodAtmMessage(int(a.Floor()), float64(takeFoodCalculation(a)))
 			a.SendMessage(-1, msg)
 			a.Log("I sent a msg", infra.Fields{"floor": a.Floor(), "message": msg.MessageType()})
 		}
 		if receivedMsg.MessageType() == "leaveFoodMessage" { //agents response to leaveFoodMessage
-			a.Log("I've been asked to leave this much food", infra.Fields{"message": receivedMsg, "Food": receivedMsg})//, infra.Fields{"foodAmt": receivedMsg.intFood})
+			a.Log("I've been asked to leave this much food", infra.Fields{"message": receivedMsg, "Food": receivedMsg}) //, infra.Fields{"foodAmt": receivedMsg.intFood})
 			//msg := *messages.NewackMessage(int(a.Floor()), true)?
 			//a.SendMessage(-1, msg)
 			a.Log("I didn't send a msg")
 		}
 		if receivedMsg.MessageType() == "takeFoodMessage" { //agents response to takeFoodMessage
-			a.Log("I've been asked to take this much food")//, infra.Fields{"foodAmt": receivedMsg.intFood})
+			a.Log("I've been asked to take this much food") //, infra.Fields{"foodAmt": receivedMsg.intFood})
 			//msg := *messages.NewackMessage(int(a.Floor()), true)?
 			//a.SendMessage(-1, msg)
 			a.Log("I didn't send a msg")
@@ -60,8 +61,7 @@ func message(a *CustomAgent3) {
 	}
 }
 
-
 //if msg != *messages.NewBaseMessage(int(a.Floor())) {
-	//	a.SendMessage(-1, msg)
-	//	a.Log("I sent a msg", infra.Fields{"floor": a.Floor(), "message": msg.MessageType()})
-	//}	
+//	a.SendMessage(-1, msg)
+//	a.Log("I sent a msg", infra.Fields{"floor": a.Floor(), "message": msg.MessageType()})
+//}
